Update Calculator.py
Layout Fixed
The default Layout is Set.
Users can't change the size of the output window.
Appearance changed. Looks more attractive.
The color codes are used for buttons so that the user can easily find the buttons even he/she is in hurry.
Padding is added.
And a few more changes.